refactor : SwipeableCarousel - Progressbar, 훅 분리
