+++ src/components/carousel/SwipeableCarousel/components/Progressbar/index.module.scss
@@ -1,0 +1,14 @@
+.progressBar {
+  margin-top: 2rem;
+  border-radius: var(--border-radius-basic);
+  background: var(--color-gray200);
+  width: 100%;
+  height: 0.4rem;
+  overflow: hidden;
+}
+
+.progressIndicator {
+  transition: width 0.3s ease;
+  background: var(--color-gray500);
+  height: 100%;
+}
