+++ src/components/carousel/SwipeableCarousel/hooks/useScrollProgress.ts
@@ -1,0 +1,22 @@
+import { useState } from 'react';
+
+interface UseScrollProgressProps {
+  containerRef: React.RefObject<HTMLDivElement | null>;
+}
+
+const useScrollProgress = ({ containerRef }: UseScrollProgressProps) => {
+  const [scrollProgress, setScrollProgress] = useState(0);
+
+  const handleScroll = () => {
+    if (!containerRef.current) return;
+    const { scrollLeft, scrollWidth, clientWidth } = containerRef.current;
+    setScrollProgress(scrollLeft / (scrollWidth - clientWidth));
+  };
+
+  return {
+    scrollProgress,
+    handleScroll,
+  };
+};
+
+export default useScrollProgress;
