+++ src/components/carousel/SwipeableCarousel/hooks/useSwipeMouse.ts
@@ -1,0 +1,45 @@
+import { useState } from 'react';
+
+interface UseSwipeMouseProps {
+  containerRef: React.RefObject<HTMLDivElement | null>;
+}
+
+const useSwipeMouse = ({ containerRef }: UseSwipeMouseProps) => {
+  const [isDragging, setIsDragging] = useState(false);
+  const [startX, setStartX] = useState(0);
+  const [scrollLeft, setScrollLeft] = useState(0);
+
+  const handleMouseDown = (e: React.MouseEvent) => {
+    if (!containerRef.current) return;
+    setIsDragging(true);
+    setStartX(e.pageX - containerRef.current.offsetLeft);
+    setScrollLeft(containerRef.current.scrollLeft);
+  };
+
+  const handleMouseMove = (e: React.MouseEvent) => {
+    if (!isDragging || !containerRef.current) return;
+
+    e.preventDefault();
+    // 드래그 속도 조절
+    const x = e.pageX - containerRef.current.offsetLeft;
+    const walk = (x - startX) * 2;
+    containerRef.current.scrollLeft = scrollLeft - walk;
+  };
+
+  const handleMouseUp = () => {
+    setIsDragging(false);
+  };
+
+  const handleMouseLeave = () => {
+    setIsDragging(false);
+  };
+
+  return {
+    handleMouseDown,
+    handleMouseMove,
+    handleMouseUp,
+    handleMouseLeave,
+  };
+};
+
+export default useSwipeMouse;
