+++ src/components/carousel/SwipeableCarousel/index.tsx
@@ -1,7 +1,9 @@
 'use client';
 
-import { useRef, useState } from 'react';
+import { useRef } from 'react';
 
+import Progressbar from './components/Progressbar';
+import { useScrollProgress, useSwipeMouse, useSwipeTouch } from './hooks';
 import styles from './index.module.scss';
 
 interface SwipeableCarouselProps {
@@ -10,57 +12,16 @@ interface SwipeableCarouselProps {
 
 const SwipeableCarousel = ({ children }: SwipeableCarouselProps) => {
   const containerRef = useRef<HTMLDivElement>(null);
-  const [isDragging, setIsDragging] = useState(false);
-  const [startX, setStartX] = useState(0);
-  const [scrollLeft, setScrollLeft] = useState(0);
-  const [scrollProgress, setScrollProgress] = useState(0);
 
-  // 🖱️ 마우스 드래그 이벤트
-  const handleMouseDown = (e: React.MouseEvent) => {
-    if (!containerRef.current) return;
-    setIsDragging(true);
-    setStartX(e.pageX - containerRef.current.offsetLeft);
-    setScrollLeft(containerRef.current.scrollLeft);
-  };
+  const { handleMouseDown, handleMouseMove, handleMouseUp, handleMouseLeave } = useSwipeMouse({
+    containerRef,
+  });
 
-  const handleMouseMove = (e: React.MouseEvent) => {
-    if (!isDragging || !containerRef.current) return;
-    e.preventDefault();
-    const x = e.pageX - containerRef.current.offsetLeft;
-    const walk = (x - startX) * 2; // 드래그 속도 조절
-    containerRef.current.scrollLeft = scrollLeft - walk;
-  };
+  const { handleTouchStart, handleTouchMove } = useSwipeTouch({
+    containerRef,
+  });
 
-  const handleMouseUp = () => {
-    setIsDragging(false);
-  };
-
-  const handleMouseLeave = () => {
-    setIsDragging(false);
-  };
-
-  // 📱 터치 이벤트 (모바일 대응)
-  const [startTouchX, setStartTouchX] = useState(0);
-  const [startScrollLeft, setStartScrollLeft] = useState(0);
-
-  const handleTouchStart = (e: React.TouchEvent) => {
-    if (!containerRef.current) return;
-    setStartTouchX(e.touches[0].pageX);
-    setStartScrollLeft(containerRef.current.scrollLeft);
-  };
-
-  const handleTouchMove = (e: React.TouchEvent) => {
-    if (!containerRef.current) return;
-    const touchX = e.touches[0].pageX;
-    const walk = (touchX - startTouchX) * 2; // 터치 드래그 속도 조절
-    containerRef.current.scrollLeft = startScrollLeft - walk;
-  };
-
-  const handleScroll = () => {
-    if (!containerRef.current) return;
-    const { scrollLeft, scrollWidth, clientWidth } = containerRef.current;
-    setScrollProgress(scrollLeft / (scrollWidth - clientWidth));
-  };
+  const { scrollProgress, handleScroll } = useScrollProgress({ containerRef });
 
   // 🎯 키보드 네비게이션 지원
   const handleKeyDown = (e: React.KeyboardEvent) => {
@@ -92,9 +53,7 @@ const SwipeableCarousel = ({ children }: SwipeableCarouselProps) => {
       >
         {children}
       </div>
-      <div className={styles.progressBar}>
-        <div className={styles.progressIndicator} style={{ width: `${scrollProgress * 100}%` }} />
-      </div>
+      <Progressbar scrollProgress={scrollProgress} />
     </div>
   );
 };
